+++ packages/utilities/index.ts
@@ -18,6 +18,7 @@ export * from "./src/JsUtils";
 export * from "./src/ProcessUtils";
 export * from "./src/TextUtils";
 export * from "./src/CliUtils";
+export * from "./src/EnvFileUtils";
 export * from "./src/doc/ISystemInfo";
 export * from "./src/diff/DiffUtils";
 export * from "./src/diff/doc/IDiffOptions";